refine code by comment
